+++ CHANGELOG.md
@@ -8,6 +8,7 @@ More detailed release notes can be found on the [releases page](https://github.c
 ### Fixed
 
 * Spaces could be appended to linux Exec command during packaging
+* Secondary mobile windows would not size correctly when padded
 
 
 ## 2.4.3 - 23 December 2023